Copy build context folder into galaxy layer
After looking at how to support offline tarballs, it might be as simple
as copying the full build context into the galaxy layer. This is okay,
becaue the layer is only used at build time.

It also allows a user to add their our data, like collection tarballs,
in the sub-folder.  Allowing for offline install of tarballs.

Signed-off-by: Paul Belanger <pabelanger@redhat.com>