Run tests on all supported run times and fix `test_build_layer_reuse` test (#326)
Run tests on all supported run times and fix `test_build_layer_reuse` test

Docker maintains a separate build cache which needs to be purged prior to running the test.
Remove the pytz fixture and instead perform the commands within each test to remove concurrency issues and make
the test easier to read.
Correctly parse the stdout from the command run. The stdout generate during the test can either be in --progress plain or --progress tty format (the default is --progress auto). I'm not sure what exactly about the environment changes the output format, so I changed the test to pass regardless of the output format.

Reviewed-by: David Shrewsbury <None>
Reviewed-by: None <None>