Support keyring and disable-gpg-verify flags of ansible-galaxy
It was requested that builder supports the opportunistic signature
verification feature of ansible-galaxy. this commit adds two new
flags. `disable-gpg-verify` lets galaxy always ignore signatures
and is simply added verbatim to the ansible-galaxy invocation in
the container file. `keyring` is the path of a GPG keyring file
that shall contain public keys that are used to verify signatures
of collections. This is forwarded to the ansible-galaxy command
by copying the given path into the build directory and passing
the path to the copied file from within the container to the
ansible-galaxy invocation