[PR ansible#731/1b659226 backport][release_3.0] 🧪 Use reusable-tox.yml @ tox-dev/workflow
This patch integrates the upstream workflow hosted at [[1]]. It
features a set of steps that are common for when having tox-centric
CI configurations. In addition to that, it also integrates automatic
verbose reruns on test failures and publishing coverage and test
results to Codecov (currently disabled).

The reusable workflow is customizable through hook points that run
in-repo actions with defined names hosted under
`.github/reusables/tox-dev/workflow/reusable-tox/hooks/`.

Additionally, an integration is added with `alls-green` [[2]] that
lets us migrate to using just a single branch protection check for the
CI instead of enumerating each possible job name in there.

The usage of `tox-uv` improved the tox env provisioning speed and
built-in caching does the same on the pip level.

PR #731

[1]: https://github.com/tox-dev/workflow
[2]: https://github.com/marketplace/actions/alls-green#why

(cherry picked from commit 1b6592267bbc0cc9a4118657faeb8680bd13ac55)