Fix introspect integration tests
Although the introspect command outputs YAML to stdout (if not
directed to write to a file), we make no guarantees about that
output being parsable YAML. In fact, it is NOT parsable YAML
since it will contain logging messages that, by default, are
surrounded with text colorization codes. Our integration tests
should not be trying to validate that output, but rather validating
the output that is written to files. The stdout output is simply
for visual validation in logging.